+++ src/python/DAS/tools/das_client.py
@@ -51,6 +51,56 @@ class DASOptionParser:
         """
         return self.parser.parse_args()
 
+def convert_time(val):
+    "Convert given timestamp into human readable format"
+    if  isinstance(val, int) or isinstance(val, float):
+        return time.strftime('%d/%b/%Y_%H:%M:%S_GMT', time.gmtime(val))
+    return val
+
+def size_format(uinput):
+    """
+    Format file size utility, it converts file size into KB, MB, GB, TB, PB units
+    """
+    try:
+        num = float(uinput)
+    except Exception as exc:
+        return uinput
+    base = 1000. # power of 10, or use 1024. for power of 2
+    for xxx in ['','KB','MB','GB','TB','PB']:
+        if  num < base:
+            return "%3.1f%s" % (num, xxx)
+        num /= base
+
+def unique_filter(rows):
+    """
+    Unique filter drop duplicate rows.
+    """
+    old_row = {}
+    row = None
+    for row in rows:
+        row_data = dict(row)
+        try:
+            del row_data['_id']
+            del row_data['das']
+            del row_data['das_id']
+            del row_data['cache_id']
+        except:
+            pass
+        old_data = dict(old_row)
+        try:
+            del old_data['_id']
+            del old_data['das']
+            del old_data['das_id']
+            del old_data['cache_id']
+        except:
+            pass
+        if  row_data == old_data:
+            continue
+        if  old_row:
+            yield old_row
+        old_row = row
+    yield row
+
 def get_value(data, filters):
     """Filter data from a row for given list of filters"""
     for ftr in filters:
@@ -59,11 +109,21 @@ def get_value(data, filters):
         row = dict(data)
         for key in ftr.split('.'):
             if  isinstance(row, dict) and row.has_key(key):
-                row = row[key]
+                if  key == 'creation_time':
+                    row = convert_time(row[key])
+                elif  key == 'size':
+                    row = size_format(row[key])
+                else:
+                    row = row[key]
             if  isinstance(row, list):
                 for item in row:
                     if  isinstance(item, dict) and item.has_key(key):
-                        row = item[key]
+                        if  key == 'creation_time':
+                            row = convert_time(item[key])
+                        elif  key == 'size':
+                            row = size_format(item[key])
+                        else:
+                            row = item[key]
                         break
         yield str(row)
 
@@ -94,7 +154,7 @@ def get_data(host, query, idx, limit, debug):
         pid = data
     else:
         pid = None
-    count = 5  # initial waiting time in seconds
+    count = 1  # initial waiting time in seconds
     timeout = 30 # final waiting time in seconds
     while pid:
         params.update({'pid':data})
@@ -154,23 +214,37 @@ def main():
             msg += ", for more results use --idx/--limit options\n"
             print msg
         mongo_query = jsondict['mongo_query']
-        if  mongo_query.has_key('filters'):
-            filters = mongo_query['filters']
+        unique = False
+        filters = mongo_query.get('filters', [])
+        aggregators = mongo_query.get('aggregators', [])
+        if  'unique' in filters:
+            unique = True
+            filters.remove('unique')
+        if  filters and not aggregators:
             data = jsondict['data']
             if  isinstance(data, dict):
                 rows = [r for r in get_value(data, filters)]
                 print ' '.join(rows)
             elif isinstance(data, list):
+                if  unique:
+                    data = unique_filter(data)
                 for row in data:
                     rows = [r for r in get_value(row, filters)]
                     print ' '.join(rows)
             else:
                 print jsondict
-        elif mongo_query.has_key('aggregators'):
+        elif aggregators:
             data = jsondict['data']
+            if  unique:
+                data = unique_filter(data)
             for row in data:
+                if  row['key'].find('size') != -1 and \
+                    row['function'] == 'sum':
+                    val = size_format(row['result']['value'])
+                else:
+                    val = row['result']['value']
                 print '%s(%s)=%s' \
-                % (row['function'], row['key'], row['result']['value'])
+                % (row['function'], row['key'], val)
         else:
             data = jsondict['data']
             if  isinstance(data, list):